Fixed test issues

+++ zkay-rsnark/src/circuit/structure/wire_ops.rs
@@ -16,7 +16,7 @@ use crate::circuit::operations::primitive::or_basic_op::{OrBasicOp, new_or};
 use crate::circuit::operations::primitive::pack_basic_op::{PackBasicOp, new_pack};
 use crate::circuit::operations::primitive::split_basic_op::{SplitBasicOp, new_split};
 use crate::circuit::operations::primitive::xor_basic_op::{XorBasicOp, new_xor};
-use crate::circuit::structure::circuit_generator::CircuitGenerator;
+use crate::circuit::structure::circuit_generator::{CircuitGenerator, getActiveCircuitGenerator};
 use crate::circuit::structure::linear_combination_wire::LinearCombinationWire;
 use crate::circuit::structure::variable_bit_wire::VariableBitWire;
 use crate::circuit::structure::variable_wire::VariableWire;
@@ -52,15 +52,15 @@ pub trait SubWire<T = Self> {
 // }
 
 pub trait XorBitwise<T = Self> {
-    fn xor_bitwise(self, w: T, numBits: i32, desc: &Option<String>) -> WireType;
+    fn xor_bitwise(self, w: T, numBits: u64, desc: &Option<String>) -> WireType;
 }
 
 pub trait AndBitwise<T = Self> {
-    fn and_bitwise(self, w: T, numBits: i32, desc: &Option<String>) -> WireType;
+    fn and_bitwise(self, w: T, numBits: u64, desc: &Option<String>) -> WireType;
 }
 
 pub trait OrBitwise<T = Self> {
-    fn or_bitwise(self, w: T, numBits: i32, desc: &Option<String>) -> WireType;
+    fn or_bitwise(self, w: T, numBits: u64, desc: &Option<String>) -> WireType;
 }
 
 pub trait IsEqualTo<T = Self> {
